// src/hooks/types.ts
export interface Point {
  type: string
  coordinates: number[]
}

export interface MultiPolygon {
  type: string
  coordinates: number[][][][]
}

export interface SensorType {
  sensor_type_id: number
  name_short: string
  name_long: string
  name_metric: string
  unit: string | null
}

export interface DataType {
  data_type_id: number
  name_short: string
}

export interface Device {
  device_id: number
  device_type_id: number
  parent_device_id: number | null
  logical: boolean
  name_short: string | null
  name_long: string | null
  capacity_dc: number | null
  capacity_ac: number | null
  point: Point | null
  polygon: MultiPolygon | null
  device_type?: DeviceType
  name_full?: string
  cec_pv_module_id?: number
  pv_module_id?: number
  device_id_path?: string
}
// ...
export interface Event {
  event_id: number
  event_type_id: number | null
  device_id: number
  device: Device
  device_name_full: string
  time_start: string
  time_end: string | null
  time_detected: string
  time_last_analyzed: string | null
  failure_mode_id: number | null
  failure_mode: FailureMode | null
  root_cause_id: number | null
  loss_total_financial: number
}
// ...
export interface FailureMode {
  failure_mode_id: number
  device_type_id: number
  name_short: string
  name_long: string
}